build: dual-license under ASL and LGPL
Add LGPL as second license, so the project can be copied into GPL2
projects. Note that you can always link to the project as-is, even
statically, from GPL2 projects. However, as soon as you actually import
the code into your project and modify it, you might end up as derivative
in the ASL-sense.

Hence, add the LGPL as second license, so people can easily import
c-list if they want to.

Signed-off-by: Tom Gundersen <teg@jklm.no>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>