c-list: add c_list_length() and c_list_contains()
These functions require to iterate the list once, and
have O(n) runtime complexity. Usually one would avoid such
operations on a list as these operations are not preferred
when using a list.

A good use for c_list_length() is to get the length of
elements to clone the list to an allocated array, like

    len = c_list_length(lst);
    arr = malloc(len * sizeof (Data));
    i = 0;
    c_list_for_each(iter, lst)
        arr[i++] = get_data(iter);

In this case, the O(n) time is fine, as the whole operation already
involves traversing the list.

c_list_contains() seems only really useful for assertion checks,
like

   remove_entry(self, entry) {
       assert(c_list_contains(self->lst_head, &entry->lst));
       // do something.
   }

Aside from an assertions, it seems that one usually would already
know whether an element is inside a list or not.

Signed-off-by: Thomas Haller <thaller@redhat.com>