c-list: make c_list_splice() clear the source
Make sure we never leave the source list stale after splicing. There is
little reason to do that. The only valid reason to not clear the source
is to save two stores. However, the source is completely invalid after
a splice, so it is unlikely to be contended. Two stores should hence
have little effect, since the source cache-line is most likely already
owned by the calling thread.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>