c-list: make c-list API const-correct
Constness in C can be useful so the compiler issues warnings
when using mutable API on supposedly immutable data.

Make all API for CList only operate on const pointers, but add
_Generic() macros that transparantly cast away constness as
appropriate.

This makes a "const CList *" pointer usable with all non-modifying
c-list API and properly preserves and propagates constness when
following the next or prev pointers. The constness also propagates
for c_list_entry(). There is c_list_entry_cast() which doesn't
exhibit this behavior.

The main point of doing this is exposing a read-only, immutable access
to traversing the CList. For example a public function

   const CList *my_object_get_xyz_list(MyObject *obj);

As the CList structues is embedded in the linked objects, that
also means that either MyObject must be a public structure or
a cast function is needed

   MyObject *my_object_from_xyz_list(const CList *);

to access the object from the list. This is also useful for internal API,
like

   static MyObject *_find_xyz(const CList *list, ...) {
       const CList *iter = ...;

       // away constness of the result??
       return c_list_entry_cast(iter, MyObject, lst);
   }

Clearly constness in C is limited and doesn't propagate well. Like
for example with _find_xyz() or strchr(). But as far as CList is
concerned, this patch properly adds implements it for the CList API.