c-list: rename c_list_unlink{,_init}
Make c_list_unlink() work like c_list_unlink_init() used to, and rename the
old c_list_unlink() to c_list_unlink_stale(). This is meant to encourage the
use of the former, and for people to only use the stale version in the extremely
rare case that it affects performance.

Signed-off-by: Tom Gundersen <teg@jklm.no>